raspberry_pi/requirements.txt
# Audio
sounddevice
# numpy
# scipy
# pyaudio
pvporcupine
pvrecorder

# VoiceVOX / Network
# requests

# Google Cloud Speech
google-cloud-speech

# Env
python-dotenv

# E-Paper
# # Pillow
# # waveshare-epaper

# Testing (dev only - not installed in production Docker)
# pytest
# pytest-mock
